+++ For42GS/For42GS/View/ContentView.swift
@@ -46,5 +46,5 @@ struct ContentView: View {
 }
 
 #Preview {
-    ContentView()
+    ContentView().environmentObject(Service())
 }
