+++ For42GS/For42GS/View/Rank/RankPageView.swift
@@ -8,10 +8,29 @@
 import SwiftUI
 
 struct RankPageView: View {
+    let items = ["level", "exp", "coalition", "comment length"]
+    @State private var selectedItem = "level" // 선택된 항목을 추적
+
+
     var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+        VStack {
+            ScrollView(.horizontal, showsIndicators: false) {
+                HStack {
+                    ForEach(items, id: \.self) { item in
+                        Text(item)
+                            .padding()
+                            .foregroundColor(self.selectedItem == item ? .orange : Color.gray) // 선택된 항목은 파란색, 그 외는 회색
+                            .onTapGesture {
+                                self.selectedItem = item
+                            }
+                    }
+                }
+                .padding()
+            }
+                RankingListView()
+            }
+        }
     }
-}
 
 #Preview {
     RankPageView()
